Backends/requirements.txt
Flask>=3.1.0
python-dotenv>=1.0.0
mysql-connector-python>=8.0.0
Flask-WTF>=1.1.1
WTForms>=3.0.0
email-validator>=1.0.0
Flask-Limiter>=2.0.0
itsdangerous>=2.1.2
Werkzeug>=3.1.0
Jinja2>=3.1.0
click>=8.0.0
gunicorn==20.1.0
